+++ AUTHORS.md
@@ -1,0 +1,7 @@
+Development Leads and Contributors
+----------------------------------
+
+* Anne Archibald (original development) <anne.archibald@nanograv.org>
+* Aaron Johnson (original fork) <aaron.johnson@nanograv.org>
+* Rutger van Haasteren (package prep and refactor) <https://github.com/vhaasteren>
+
