+++ README.md
@@ -1,2 +1,26 @@
 # h5pulsar
- Fork of Anne Archibald's PR to add hdf5 to Pulsar objects
+
+Allow Enterprise pulsars to be written to and read from HDF5 files. Especially
+when timing packages like Tempo2 or PINT are either not available or very slow,
+this can be a big aid.
+
+# Installation
+
+The h5pulsar add-on to Enterprise can be installed straight from github using
+
+```bash
+pip install git+https://github.com/vhaasteren/h5pulsar.git
+```
+
+# From Pypi, you can do:
+NOTE: not uploaded to Pypi yet
+
+```bash
+pip install h5pulsar
+```
+
+# Conda
+
+```bash
+conda install h5pulsar
+```
