+++ requirements.txt
@@ -1,8 +1,5 @@
-astropy >=5.3
-enterprise_pulsar >=3.3.2
-h5py >=3.8.0
-libstempo >=2.4.5
-numpy >=1.24.3
-pint_pulsar >=0.9.5
-packaging >=23.1
-ephem >=4.1.4
+numpy>=1.16.3
+astropy>=4.0,!=4.0.1,!=4.0.1.post1
+enterprise-pulsar>=3.3.4
+jplephem>=2.6
+h5py>=3.9.0
